backend/requirements.txt
# Core
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment
python-decouple==3.8

# Auth
djangorestframework-simplejwt==5.3.1
django-allauth==0.63.3
dj-rest-auth==6.0.0
cryptography==47.0.0

# Cloudinary
cloudinary==1.40.0
django-cloudinary-storage==0.3.0

# Email / Async
# Using threading for simplicity (no Celery needed)

# Rate limiting
django-ratelimit==4.1.0

# Static files
whitenoise==6.7.0

# Gunicorn
gunicorn==22.0.0

# File export
openpyxl==3.1.3
reportlab==4.2.2

# Utilities
Pillow==10.3.0
django-filter==24.2
